Include real modules directly in Sphinx doc
Cross-module references have been updated accordingly.  In most cases,
all that was needed was to put at "." at the beginning of the reference
and ensure it searched for the right kind of item (class/meth/func/
data).

We can now directly document module-level variables and have them show
up in the Sphinx output.